Issue #23: Use template redirects on DBpedia server
With this change the Server will read template redirects information from the
Wiki statistics as extracted by the CreateStats script.
The Server only polishes the data a bit in order to remove the template namespace
from the WikiTitles. Redirects are then propagated to ExtractionManagers which add them
to the Extractors contexts.