Select correct MIME type for server sample extraction response
Currently the Server always responds with a "application/xml" MIME type
when trying to extract triples from a Wikipage, no matter which format
is selected in the form.

When it comes to formats which are not TriX, the rendering on modern
browser fails due to incorrect files w.r.t. the declared MIME type of the
HTTP response.

With this pull request, the "application/xml" MIME type is selected only if
TriX format is picked, otherwise "text/plain" is used.