Issue #106: builders for template property value
There are cases in which it is more appropriate to assign pre-computed
values to template properties mapping instead of reading them from the
actual template used in a Wikipedia articles.

E.g. Infobox_Australian_place let editors specify lat/long values for geopoints
but hardcodes directions (latDir and longDir) in the template definition.
Since the property value is built using GeoCoordinateMapping it is impossible
to define a constant value for latDir and longDir with the current language.
(In the case of SimplePropertyMappings it is still possible to use ConstantMapping).

With this commit an extension of the framework is provided to let editors
have better control of how a template property is mapped into an ontology property.

Future enhancements of this approach could be:
- selection in a list of values (currently the only way to do that is in SimplePropertyMapping
using the select property)
- prefix/suffix handling with template property value
- internal/external link generation from string values (e.g. imdbId to ExternalLink("http://imdb.com/id/imdbId"))