Better support for URL template transformation
The URL template in Wikipedia allows editors to insert also
a non-absolute URL, resolving it for them with a default URI scheme.

There are some cases in Wikipedia articles where this applies, and the
framework should support this.

With this patch Template:URL is transformed adding a default http:// URI scheme
if the URI does not provide one.

Also some unit tests have been added for the LinkParser