Fix HomepageExtractor issues
The HomepageExtractor was not able to correctly detect references to official webpages
when conveyed through templates placed in the External links section.
Only matches to Template:Official were considered and the URL extraction was failing
when the template property contained an ExternalLinkNode.

With this change:
- A map is provided to store info about official website templates, for different langs
- Template redirects are checked when looking for official websites
- URL is extracted also when the template property is an ExternalLinkNode

I was able to extract 550k homepages from enwiki-20130604 compared to about 505k items
without this patch (about 10% more).