HomepageExtractor config
Set default values in case any of the HomepageExtractor required config does not exist
for a specific language (e.g. there is no official website template in itwiki, but
there is an official keyword used in external links section).

Not using English as default because it looks like a strong assumption which is likely
to be either wrong or disruptive.