Do not capitalize page titles when importing dump
Capitalization of titles led to duplicates in SQL key for
namespace and title if a corresponding upper-case title was
already contained in the dump.

WikiTitle now has a new method parseCleanTitle which can be applied to
clean titles like those contained in Wikipedia dumps. Capitilization can
now also be disabled in the constructor of WikiTitle.